backend/requirements.txt
fastapi
uvicorn
python-multipart
easyocr
groq
supabase
pydantic
pydantic-settings
python-dotenv
axios
slowapi
python-jose[cryptography]
pillow
numpy
pandas
